requirements.txt
openai
ffmpeg-python
openai-whisper
streamlit
pytube